extract filter builder out of filter widget
was getting too messy
use formlayout instead of gridlayout for filter builder
move search_classes to filter.cpp and rename to filter_classes
